mimir: Rules loading no longer hits the disk (#13156)
#### What this PR does

This change is an extension of
https://github.com/grafana/mimir/pull/12617 which put all of the ruler's
filesystem interactions behind an abstraction.

The ruler's Prometheus components expect rules to be passed to them via
disk, and the APIs mostly accept rule file paths. To support this, the
ruler normally mirrors the set of files to load to disk and then tells
Prometheus ruler managers to load them back in - effectively using the
disk as a communication method between two portions of the software in
the same process.

This PR changes the filesystem to a virtual one. The rule files get
loaded like normal, but stay in memory.

#### Which issue(s) this PR fixes or relates to

Fixes #<issue number>

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.