parquet store-gateway: track all blocks, check for converted on demand (#13179)
With the current approach, we only check for conversion marks during
block discovery, which happens (by default) every 15 mins. This can lead
to a scenario where get a query for a block has already been converted
but we haven't ran a discovery since then -- we don't query it,
resulting in a consistency check error from queriers.

This makes it so we track blocks as soon as we see them even if they are
not converted, and check for conversion markers on the fly during a
query. "true" results for conversion checks are cached forever inmemory,
but for "false" results we cache the result for a minute to prevent
useless spikes of `Exists()` calls.