compactor scheduler: implement planning (#13153)
This implements planning compaction jobs for the compaction worker. If
you would like to follow along with the flow the idea is to do what the
set up in `compactUser` does in `MultitenantCompactor`, but with these
goals:
- provide access to creating the `MetaSyncer`(to prepare for possible
future remote caching)
- provide access to creating the `BucketCompactor` (to help with the
implementation of compaction)
- not filter for owned jobs
- refactor the existing flow as little as possible to reduce potential
impact

I have not added unit tests for the actual planning. I think it would be
valuable to do so since there's some code duplication here, but I'd also
like to make this available for the compaction implementation sooner
rather than later.

As a note for the unit test changes, `c.shardingStrategy` had to be
created due to `c.shardingStrategy.ownJob` being passed during
`newBucketCompactor` (it isn't used by the worker). The other changes
are similar in nature where it's about getting the mock set up correctly
to not fail when the planning executes.