mirrored querier: Use primary query finished time to set cutoff for ignoring chunks (#13185)
<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does

We discard series from the secondary querier based on some timestamp.
Previously this timestamp was before our first request started. This
lead to discarding some series from the secondary querier, which were
created during the call to the primary querier and which the primary
querier returned.

This PR fixes that by ignoring only series created since the first
querier returned (again - based on sample timestamps, which isn't a
guarantee)


#### Which issue(s) this PR fixes or relates to

Fixes #<issue number>

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Co-authored-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>