Tooling: Allow renovate to update regular dependencies daily (#13210)
#### What this PR does
This PR amends the general renovate schedule to use the Daily option (`*
0-3 * * *`), and adds a package rule to limit digest updates to once
weekly (`* 0-4 * * 1`) as we have several dependencies that just use
digests/hashes on a repo's main branch instead of semver releases.

#### Which issue(s) this PR fixes or relates to

Fixes #<issue number>

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.