Remote execution: fix issue where evaluation metrics and logs aren't emitted (#13207)
#### What this PR does

This PR fixes an issue where evaluation metrics and logs aren't emitted
by queriers if remote execution is enabled.

This happened because these metrics and logs were emitted by
`Query.Exec()`, but `Query.Exec()` is only invoked in query-frontends,
not queriers, if remote execution is enabled.

#### Which issue(s) this PR fixes or relates to

https://github.com/grafana/mimir/pull/12302

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.