Dashboards: Update default regular expressions to match multi-zone deployments for query-frontend, querier, distributor and ruler (#13200)
#### What this PR does

I'm working on multi-zone support for read-path components too. In this
I'm proposing to update the default regular expressions – used to
generate dashboards and some alerts - to match multi-zone deployments
for query-frontend, querier, distributor and ruler.

#### Which issue(s) this PR fixes or relates to

N/A

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Marco Pracucci <marco@pracucci.com>