feat: remove experimental heartbeat/timeout disabling (#13104)
## What

Removes support for setting heartbeat period or timeout to 0 to disable
them.

## Why

This experimental feature was added in
https://github.com/cortexproject/cortex/pull/4342 but is not used in
production. Removing unused experimental features simplifies the
codebase.

## Changes

- Remove "Hash ring" section from experimental features documentation
listing the heartbeat disabling flags
- Add CHANGELOG entry under `[CHANGE]` scope
- Update dskit dependency to include validation changes (via replace
directive for now)

## Dependencies

This PR depends on https://github.com/grafana/dskit/pull/759 which must
be merged first. After that PR is merged, this PR will be updated to:
1. Remove the replace directive
2. Update dskit to the merged commit
3. Update the CHANGELOG with the actual PR number