Remote execution: fix issue where subqueries can fail with `slice capacity must be a power of two, but is X` (#13211)
#### What this PR does

This PR fixes an issue where subqueries evaluated as the root node of a
remote execution request can fail with `slice capacity must be a power
of two, but is X`.

MQE's ring buffer implementations require that the slices used have
capacities that are powers of two.

We end up with a slice with capacity that isn't a power of two because
the slices created by unmarshalling remote execution result messages
aren't drawn from the usual pools that MQE uses, and so Go's usual
behaviour around slice capacity when appending items applies: if the
slice length is less than 1024, then the capacity is doubled, otherwise
it is increased by 25%, so there's no guarantee it's a power of two.

I've applied the same change for instant vector and scalar results to
ensure that these slices don't cause problems if they are later used as
the result of a subquery.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

(cherry picked from commit d97c025c791f0796066106fc9dfb745e53471604)