Remote execution: log errors returned by queriers (#13246)
#### What this PR does

This PR changes the behaviour to log any errors returned by queriers, in
addition to attaching an event to the active trace span.

I've chosen not to add a changelog entry given this is a minor logging
change not visible to users.

#### Which issue(s) this PR fixes or relates to

https://github.com/grafana/mimir/pull/12302

#### Checklist

- [n/a] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.