feat(query-tee): Include /api/v1/read as an existing route (#13227)
The current query tee did not have api/v1/read registered despite of it
being both a valid Mimir and Prometheus route. Its not problematic if
`PassThroughNonRegisteredRoutes` is set to true. However with the new
cross cluster validation header requirement, we need the route to be
registered so headers get added to reverse proxied requests.

---------

Signed-off-by: Jesus Vazquez <jesus.vazquez@grafana.com>