GitHub: Let build and push workflow execute for all PRs (#13225)
<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does

Let the build and push GitHub workflow execute for all PRs, so it can be
made a required check (so it's not skipped before merging).

Context: I think https://github.com/grafana/mimir/pull/13212 got
auto-merged (by me) before the image finished building, because it's not
a required check.

I'm also removing the `on.pull_request.types` constraint, to harmonize
with the `ci` workflow, just to make sure it can be a required check on
all PRs.

#### Which issue(s) this PR fixes or relates to

Fixes #<issue number>

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Arve Knudsen <arve.knudsen@gmail.com>