parquet converter: pass lookback to MetaFetcher for faster discovery (#13257)
#### What this PR does

We are currently fetching _all_ Metas (except for the blocks marked for
deletion) and postfiltering them based on our criteria. That
post-filtering happens after the Metas are parsed, which is expensive.
One of our filters, `min_block_timestamp`, can be translated into the
MetaFetcher's "lookback" filter, which is applied on the blocks ULIDs
before Metas are parsed. This should speed up discovery if such filter
is being used, and it filters a lot.

Note that the MetaFetcher takes the lookback and does
`time.Now().Add(-f.maxLookback)`, while our `min_block_timestamp` is
absolute. So we are sensitive to the time at which the MetaFetcher
actually does that. To compensate for that, I'm adding an hour to the
lookback, which should be more than enough in practice, and we still
post-filter the same way. For converters, we don't really care about
these filters being _exact_ anyway.

#### Which issue(s) this PR fixes or relates to

n/a

Fixes #<issue number>

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.