distributor: Uncover cleanup races in StartFinishRequest test (#13269)
#### What this PR does

After https://github.com/grafana/mimir/pull/13003, we had a race
condition between the `reactiveLimiterCleanup`, which reads
`requestState.pushErr`, and the defer in `limitsMiddleware` that sets
`requestState.pushErr`.

The race happens because push request cleanup runs concurrently with the
parent goroutine, that handles the request and eventually calls the
defer. (IMHO this is in itself an objectionable choice.)

This wasn't caught by the race detector in tests, because the tests
weren't running cleanup funcs concurrently. Now they do, and they do
detect the race.

The race condition is fixed by running `reactiveLimiterCleanup` in the
defer, called synchronously, instead of the push cleanups, called in a
separate goroutine.

In the meantime, https://github.com/grafana/mimir/pull/13239 fixed the
race condition in another way. I only noticed this after doing my own
fix, but I'm still submitting it for consideration, mostly because I
think it's preferable the cleanup func isn't called concurrently.

In any case, we can also just merge
8130f9f4d26934ec07fc0ffef40ab0f015606bf4, to try and avoid new code from
introducing race conditions again.

#### Which issue(s) this PR fixes or relates to

* https://github.com/grafana/mimir/pull/13003
* https://github.com/grafana/mimir/pull/13239

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.