ruler: Add tests for ruler.mapper (#13265)
#### What this PR does

In https://github.com/grafana/mimir/pull/13264 we intend to replace the
`mapper` with another component that shares a similar interface.

Before doing so, the `mapper` does not have very deep tests. Let's add
more tests to fully characterize its behavior, so we can confidently
make the switch.

The new tests specifically focus on:
- multi-tenancy and behavior when multiple tenants are mapped
simultaneously
- asserting on the contents of the stored rules and proving they can be
read back out of the filesystem

#### Which issue(s) this PR fixes or relates to

n/a

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.