store-gateway: fix TestBucketChunkReader_loadChunks_verifiesCRC32 (#13258)
follow-up on #13151; thanks @jhalterman for reporting this

#### Problem

The `naivePartitioner` that we were using created a part for each
object. The concurrency in fetching multiple parts from object store
made the order of fetching chunks unpredictable. So randomly different
chunks would be checksummed. This broke the assumption in the test code.

#### What this PR does

* use gap-based partitioners which effectively merge the whole range
into one part that we fetch
* changed `createChunkData` to be a bit more predictable. It was hard to
debug the chunk contents if they're all random

<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->


#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.