usage-tracker: improve snapshot loading & rehash performance
When we're loading the first snapshot and when we're rehashing, we know that the keys we're adding do not exist in the tenantshard yet, so checking their existence is a waste of time.

This adds a new tenantshard.Map.Load() method that skips the lookup, and adds the element to the first empty slot available.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>