mixin: update MimirHighVolumeLevel1BlocksQueried to fire on percentage of L1 blocks (#13229)
#### What this PR does

This is the follow-up to https://github.com/grafana/mimir/pull/11803

This PR updates the `MimirHighVolumeLevel1BlocksQueried` alert to fire
when the percentage of the queried out-of-order level 1 blocks exceed a
threshold.

Inside the GL production we currently observe the previous version of
the alert. I haven't got to roots of why this happens. But I DONT see
any evidence of compactors being the problem. My theory is that there is
a tiny percentage of users, with a pathological workload.

E.g. I think, it would be enough to have one user in a cell, that
doesn't produce samples regularly, but has a recording rule (or an
alert), that runs every N hours, and queries the data from the past 12+
hours.

I've back-tested the updated query against the data from June, when we
had an incident with the compactors. The updated query would find the
problematic cell

<img width="1454" height="381" alt="Screenshot 2025-10-29 at 14 13 50"
src="https://github.com/user-attachments/assets/c83347ea-7b13-453e-8f35-e96601fe7de8"
/>

---------

Signed-off-by: Vladimir Varankin <vladimir.varankin@grafana.com>