Small benchmark improvements (#13299)
#### What this PR does

Trying to iron out a couple of things that slowed me up when starting to
use this tool.

* StreamingPromQL Benchmarks: log errors to stderr - if something goes
wrong during the benchmark, e.g. a WAL file is
corrupt, it's useful to have that printed out.
* Also it's now easy to edit the line from `"warn"` to `"info"` so you
can see progress, which is nice when loading gigabytes of data.
* Improve README
  * Subdirectory is hard-coded to `data`
* Add `--forward` -- it's more logical to execute queries in the same
order they happened originally.

#### Question
Should the logger be set to `"error"` level? There is one warning that
may come out frequently:
```
level=warn msg="-blocks-storage.backend=filesystem is for development and testing only; you should switch to an external object store for production use or use a shared filesystem"
```

#### Checklist

- [x] Tests updated.
- NA Documentation added.
- NA `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`,
`[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry is not
needed, please add the `changelog-not-needed` label to the PR.
- NA
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.