helm: correct nginx removal version from 7.0 to 6.0 (#13294)
The nginx section was deprecated in 4.0.0 with the intention to remove
it in 7.0.0, but it was actually removed in 6.0.0.

Updated the deprecation notice in the CHANGELOG and added an addendum to
the v4.0 release notes to clarify this.

(cherry picked from commit d5455b86566ee37e35193a397824cef1ba53a226)