feat(query-tee): Add cluster validation configuration (#13302)
#### What this PR does

This PR reuses dskit clustervalidation configuration to make it possible
for query-tee to set the X-Cluster header if the user wants to. Before
it was possible to do this by setting X-Cluster header in the backend
configuration but I figured this would be easier if all backends belong
to the same cluster.


#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Jesus Vazquez <jesus.vazquez@grafana.com>
Co-authored-by: Arve Knudsen <arve.knudsen@gmail.com>