code review: change shortcut for skipping matcher reduction
Change the check for when to skip trying to reduce matchers. We
should only attempt to reduce matchers when there is more than one
matcher. This also requires removing a test that made sure a single
wildcard matcher wasn't removed because a single wildcard matcher
is not a valid query.

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>