docs: Clarify that distributor runs in both ruler operational modes (#13300)
#### What this PR does

The current docs mention the internal distributor under the "Operational
modes" > "Internal" section. But it should apply to both operational
modes, as (as far as I can tell) both use an internal distributor in the
same way.


https://github.com/grafana/mimir/blob/5de0cf555010f0f26a35518199dfeae16888f6b2/pkg/mimir/modules.go#L1142-L1144

#### Which issue(s) this PR fixes or relates to

—

#### Checklist

- [ ] Tests updated.
- [x] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Co-authored-by: Arve Knudsen <arve.knudsen@gmail.com>