[cost attribution]: get rid of remaining MustNewConstMetric calls (#13282)
#### What this PR does

This PR completes the refactoring of cost attribution metrics by
replacing the remaining prometheus.MustNewConstMetric calls in the cost
attribution manager with the new descriptor type introduced in #13273

#### Which issue(s) this PR fixes or relates to

Fixes #<issue number>

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

Signed-off-by: Yuri Nikolic <durica.nikolic@grafana.com>