Mimir-rules-action needs base alpine image (#13303)
<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does

Following https://github.com/grafana/mimir/pull/13014, which changed the
base mimirtool image to be distroless, the mimir-rules-action Github
Action has failed to run with error `exec /entrypoint.sh: no such file
or directory`.

This PR returns the mimir-rules-action to be based on an alpine linux
base image, and copies in the mimirtool binary.

#### Which issue(s) this PR fixes or relates to

Fixes https://github.com/grafana/mimir/issues/13263

#### Checklist

- NA Tests updated.
- NA Documentation added.
- [X] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- NA
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Vladimir Varankin <vladimir.varankin@grafana.com>
Co-authored-by: Vladimir Varankin <vladimir.varankin@grafana.com>