alertmanager: Deprecate UTF-8 strict mode flag and default to true (#13109)
Changes the `-alertmanager.utf8-strict-mode-enabled` flag to default to
`true` and marks it as deprecated. The flag was first introduced as
experimental in Mimir 2.12, promoted to stable in Mimir 3.1, and is now
deprecated. It will be removed in Mimir 3.3, at which point UTF-8 strict
mode will be the only supported behavior.