Fix flaky TestQuerierTenantFederationWithRemoteExecution (#13352)
#### What this PR does

This PR fixes an issue where the
`TestQuerierTenantFederationWithRemoteExecution` integration test would
sometimes flake due to queries arriving at the query-frontend before the
query-frontend had a complete view of the querier ring.

#### Which issue(s) this PR fixes or relates to

#12757 

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.