query-tee: fix route registration now that we're using dskit config (#13360)
#### What this PR does

In https://github.com/grafana/mimir/pull/13328 we introduced dskit flags
and in https://github.com/grafana/mimir/pull/13355 we fixed the double
registration of log level and path prefix. In this PR we are now fixing
routing registration for query tee.

dskit server automatically strips out the prefix path using
[PathPrefix().Subrouter()](https://github.com/grafana/dskit/blob/main/server/server.go#L398-L402)
but because we were registering the routes with the prefix then we were
getting lots of 404s.

i.e
- Request to /prometheus/api/v1/query with
server.path-prefix=/prometheus
- Before: Route was registered as /prometheus/api/v1/query but because
dskit was removing the prefix from incoming requests with the subrouter
functionality it was leading to 404s
- Now: Route registered as /api/v1/query and that works out OK with
dskit subrouter

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.