mixin: plot OOMKilled events in resources dashboards (#13377)
#### What this PR does

The PR updates the dashboards to plot the per-pod OOMKilled events on
the memory's working set panels. This is how it looks like, when viewing
one of our existing clusters (note red dots 🔴 on the central panel)

<img width="1456" height="306" alt="Screenshot 2025-11-06 at 12 50 42"
src="https://github.com/user-attachments/assets/f153bf3f-4f3f-4234-ad95-347b0109c3dd"
/>

To start, I'm only updating the dashboards for the Kubernetes deployment
types. We can later iterate on the implementation, if someone knows how
to do that for bare-metal.

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Vladimir Varankin <vladimir.varankin@grafana.com>